+++ src/readme.test.ts
@@ -1,4 +1,4 @@
-import { pipe, pip, pipeA, piper, piperA } from "./index";
+import { pipe, pip, pipeA } from "./index";
 import * as fp from "lodash/fp";
 import * as _ from "lodash";
 
@@ -22,12 +22,16 @@ const marbles: Marble[] = [
 const user = { marbles };
 
 const isRed = (marble: Marble) => marble.color === "red";
-const filterReds = fp.filter(isRed);
+const onlyRed = fp.filter(isRed);
+const isSmall = (marble: Marble) => marble.size === 1;
+const onlySmall = fp.filter(isSmall);
 const getLength = <T>(xs: T[]): number => xs.length;
 
 const fetchMarbles = (user: User) => Promise.resolve(user.marbles);
+const fetchMarblesWillFail = (_user: User) =>
+  Promise.reject("you've lost your marbles!");
 const fetchFavoriteColor = () => Promise.resolve("red");
-const filterWithAsyncColor = async (marbles: Marble[]) => {
+const filterForFavoriteColor = async (marbles: Marble[]) => {
   const color = await fetchFavoriteColor();
   return _.filter(marbles, marble => marble.color === color);
 };
@@ -35,7 +39,10 @@ const filterWithAsyncColor = async (marbles: Marble[]) => {
 describe("readme", () => {
   describe("pipe", () => {
     it("handles the readme example", () => {
-      const redCount = pipe(marbles)(filterReds)(getLength).value;
+      const redCount = pipe(marbles)
+        .thru(onlyRed)
+        .thru(getLength)
+        .value();
 
       expect(redCount).toBe(3);
     });
@@ -43,47 +50,81 @@ describe("readme", () => {
 
   describe("pip", () => {
     it("handles the readme example", () => {
-      const redCount = pip(marbles)(filterReds)(getLength).value;
+      const redCount = pip(marbles)
+        .thru(onlyRed)
+        .thru(getLength)
+        .value();
 
       expect(redCount).toBe(3);
     });
   });
 
-  describe("piper", () => {
+  describe("point-free pipe", () => {
     it("handles the readme example", () => {
-      const redCounter = piper(filterReds)(getLength).run;
+      const redCounter = pipe.thru(onlyRed).thru(getLength);
       const redCount = redCounter(marbles);
 
       expect(redCount).toBe(3);
+    });
+
+    it("handles the readme example about immutable updates", () => {
+      const getSmallReds = pipe.thru(onlyRed).thru(onlySmall);
+      const smallRedCounter = getSmallReds.thru(getLength);
+
+      const smallReds = getSmallReds(marbles);
+      const smallRedCount = smallRedCounter(marbles);
+
+      expect(smallReds.length).toBe(1);
+      expect(smallRedCount).toBe(1);
     });
   });
 
   describe("pipeA", () => {
     it("handles the readme example", async () => {
-      // prettier-ignore
-      const redCount = await pipeA
-        (user)
-        (fetchMarbles)
-        (filterWithAsyncColor)
-        (getLength)
-        .value;
+      const redCount = await pipeA(user)
+        .thru(fetchMarbles)
+        .thru(filterForFavoriteColor)
+        .thru(getLength)
+        .value();
 
       expect(redCount).toBe(3);
     });
+
+    it("handles the readme example about errors", async () => {
+      const redCount = await pipeA(user)
+        .thru(fetchMarblesWillFail)
+        .thru(filterForFavoriteColor)
+        .thru(getLength)
+        .value()
+        .catch(() => 0);
+
+      expect(redCount).toBe(0);
+    });
   });
 
-  describe("piperA", () => {
+  describe("point-free pipeA", () => {
     it("handles the readme example", async () => {
-      // prettier-ignore
-      const redCounter = piperA
-        (fetchMarbles)
-        (filterWithAsyncColor)
-        (getLength)
-        .run;
+      const redCounter = pipeA
+        .thru(fetchMarbles)
+        .thru(filterForFavoriteColor)
+        .thru(getLength);
 
-      const redCount = await redCounter(user);
+      const redCount = await redCounter(Promise.resolve(user));
 
       expect(redCount).toBe(3);
+    });
+
+    it("handles the readme example about errors", async () => {
+      const redCounter = pipeA
+        .thru(fetchMarblesWillFail)
+        .thru(filterForFavoriteColor)
+        .thru(getLength);
+
+      try {
+        await redCounter(Promise.resolve(user));
+      } catch (error) {
+        expect(error).toBe("you've lost your marbles!");
+      }
     });
   });
 });